+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - README now contains links to mentioned tools
  - Log node version on startup.
  - Give meaningful error when required auth user is not defined.
+ - Add option to allow bypassing auth for configurable IP ranges.
 
 # 0.6.0